requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
python-docx==0.8.11
Pillow==10.0.0
gunicorn==21.2.0